CS4182/assignment_code/requirements.txt
PyOpenGL-3.1.5-cp27-cp27m-win_amd64.whl
Django==1.11.29
image==1.5.32
Pillow==6.2.2
pytz==2020.1
six==1.15.0
numpy
